apps: split scaler client
